Package Arrangement and clean up code
