+++ app/src/main/java/com/singularity/ipcaplus/chat/ChatsFragment.kt
@@ -1,42 +1,36 @@
-package com.singularity.ipcaplus.home
+package com.singularity.ipcaplus.chat
 
-import android.content.ContentValues
 import android.content.Intent
-import android.graphics.Color
-import android.graphics.drawable.Drawable
+import android.os.Build
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.appcompat.content.res.AppCompatResources
-import androidx.core.graphics.drawable.DrawableCompat
+import androidx.annotation.RequiresApi
 import androidx.recyclerview.widget.DefaultItemAnimator
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
-import com.singularity.ipcaplus.Backend
-import com.singularity.ipcaplus.ChatActivity
+import com.singularity.ipcaplus.utils.Backend
 import com.singularity.ipcaplus.R
-import com.singularity.ipcaplus.Utilis
+import com.singularity.ipcaplus.utils.Utilis
+import com.singularity.ipcaplus.cryptography.saveKeygenOx
 import com.singularity.ipcaplus.databinding.FragmentChatsBinding
-import com.singularity.ipcaplus.databinding.FragmentOfficialChatsBinding
 import com.singularity.ipcaplus.models.Chat
 
 
-class OfficialChatsFragment : Fragment() {
+class ChatsFragment : Fragment() {
 
     // Variables
     var chats = arrayListOf<Chat>()
     var chatIds = arrayListOf<String>()
-    var user_groups = arrayListOf<String>()
 
-    private var _binding: FragmentOfficialChatsBinding? = null
+    private var _binding: FragmentChatsBinding? = null
     private var mAdapter: RecyclerView.Adapter<*>? = null
     private var mLayoutManager: LinearLayoutManager? = null
 
@@ -48,10 +42,12 @@ class OfficialChatsFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentOfficialChatsBinding.inflate(layoutInflater)
+
+
+        _binding = FragmentChatsBinding.inflate(layoutInflater)
         val root: View = binding.root
 
-        // Get Oficial Chats
+        // Get Group Chats
         db.collection("profile").document("${Firebase.auth.currentUser!!.uid}").collection("chat")
             .addSnapshotListener { documents, e ->
                 documents?.let {
@@ -59,7 +55,7 @@ class OfficialChatsFragment : Fragment() {
                     chatIds.clear()
                     for (document in it) {
                         val chat = Chat.fromHash(document)
-                        if (chat.type == "oficial") {
+                        if (chat.type == "chat") {
                             chats.add(chat)
                             chatIds.add(document.id)
                         }
@@ -68,13 +64,12 @@ class OfficialChatsFragment : Fragment() {
                 }
             }
 
-        // RecyclerView Chats
+        // RecyclerView Chat
         mLayoutManager = LinearLayoutManager(activity, LinearLayoutManager.VERTICAL, false)
         binding.recyclerViewGroups.layoutManager = mLayoutManager
         mAdapter = ChatAdapter()
         binding.recyclerViewGroups.itemAnimator = DefaultItemAnimator()
         binding.recyclerViewGroups.adapter = mAdapter
-
 
         return root
     }
@@ -84,25 +79,18 @@ class OfficialChatsFragment : Fragment() {
         _binding = null
     }
 
-    /*
-       This function configures the fragment after its creation
-    */
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-    }
-
-
-      inner class ChatAdapter : RecyclerView.Adapter<ChatAdapter.ViewHolder>() {
+    inner class ChatAdapter : RecyclerView.Adapter<ChatAdapter.ViewHolder>() {
 
         inner class ViewHolder(val v: View) : RecyclerView.ViewHolder(v)
 
         override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-            return ViewHolder(
-                LayoutInflater.from(parent.context).inflate(R.layout.row_chat, parent, false)
-            )
+                return ViewHolder(
+                    LayoutInflater.from(parent.context).inflate(R.layout.row_chat, parent, false)
+                )
+
         }
 
+        @RequiresApi(Build.VERSION_CODES.M)
         override fun onBindViewHolder(holder: ViewHolder, position: Int) {
 
 
@@ -114,7 +102,11 @@ class OfficialChatsFragment : Fragment() {
                     val lastMessageTime = findViewById<TextView>(R.id.lastMessageTime)
                     val lastMessageText = findViewById<TextView>(R.id.textViewLastMessage)
 
+
                     textViewMessage.text = chats[position].name
+
+                    // sync data recieved form direbase with encrypted shared preferences (key -> 1x)
+                    saveKeygenOx(context, chatIds[position], chats[position].ox.toString())
                     // Set Last Chat Message
                     Backend.getLastMessageByChatID(chatIds[position]) {
                         val data = Utilis.getDate(it!!.time.seconds *1000, "yyyy-MM-dd'T'HH:mm:ss.SSS")
@@ -132,17 +124,18 @@ class OfficialChatsFragment : Fragment() {
 
         }
 
-          override fun getItemViewType(position: Int) : Int {
-              if (chats[position].type == "oficial") {
-                  return 1
-              } else {
-                  return 0
-              }
-          }
+        override fun getItemViewType(position: Int) : Int {
+            if (chats[position].type == "chat") {
+                return 1
+            } else {
+                return 0
+            }
+        }
 
         override fun getItemCount(): Int {
             return chats.size
         }
     }
+
 
 }
