+++ app/src/main/java/com/singularity/ipcaplus/chat/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.singularity.ipcaplus.home
+package com.singularity.ipcaplus.chat
 
 import android.os.Bundle
 import android.view.LayoutInflater
