+++ app/src/main/java/com/singularity/ipcaplus/drawer/DateTimePicker.kt
@@ -1,4 +1,4 @@
-package com.singularity.ipcaplus.calendar
+package com.singularity.ipcaplus.drawer
 
 import android.content.Context
 import android.graphics.Canvas
