+++ app/src/main/res/layout/activity_add_event.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".calendar.AddEventActivity">
+    tools:context=".drawer.AddEventActivity">
 
     <TextView
         android:id="@+id/textViewDateTime"
@@ -20,7 +20,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.singularity.ipcaplus.calendar.DateTimePicker
+    <com.singularity.ipcaplus.drawer.DateTimePicker
         android:id="@+id/dateTimePicker2"
         android:layout_width="0dp"
         android:layout_height="150dp"
