+++ app/src/main/res/layout/activity_calendar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".calendar.CalendarActivity">
+    tools:context=".drawer.CalendarActivity">
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/nestedScrollView"
