+++ app/src/main/res/layout/activity_chat.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingRight="10dp"
-    tools:context=".ChatActivity">
+    tools:context=".chat.ChatActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycleViewChat"
