+++ app/src/main/res/layout/activity_chat_more.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ChatMoreActivity">
+    tools:context=".chat.ChatMoreActivity">
 
     <ImageView
         android:id="@+id/imageViewGroup"
