+++ app/src/main/res/layout/activity_contacts.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".profile.ContactsActivity">
+    tools:context=".drawer.ContactsActivity">
 
     <ExpandableListView
         android:id="@+id/expandableListViewContacts"
